Extract common manifest validators
