+++ datadog_checks_dev/datadog_checks/dev/tooling/commands/validate/manifest.py
@@ -9,6 +9,7 @@ import click
 from datadog_checks.dev.tooling.manifest_validator import get_all_validators
 
 from ....fs import file_exists, read_file, write_file
+from ....structures import JSONDict
 from ...constants import get_root
 from ...testing import process_checks_option
 from ...utils import complete_valid_checks
@@ -48,6 +49,7 @@ def manifest(ctx, check, fix):
 
             try:
                 decoded = json.loads(read_file(manifest_file).strip())
+                decoded = JSONDict(decoded)
             except json.JSONDecodeError as e:
                 failed_checks += 1
                 echo_info(f"{check_name}/manifest.json... ", nl=False)
@@ -59,6 +61,9 @@ def manifest(ctx, check, fix):
             all_validators = get_all_validators(ctx, version, is_extras, is_marketplace)
 
             for validator in all_validators:
+                if validator.skip_if_errors and file_failures > 0:
+                    echo_info(f'Skipping validation {validator} since errors have already been found.')
+                    continue
                 validator.validate(check_name, decoded, fix)
                 file_failures += 1 if validator.result.failed else 0
                 file_fixed += 1 if validator.result.fixed else 0
