+++ datadog_checks_dev/datadog_checks/dev/tooling/manifest_validator/common/__init__.py
@@ -1,0 +1,3 @@
+#  (C) Datadog, Inc. 2021-present
+#  All rights reserved
+#  Licensed under a 3-clause BSD style license (see LICENSE)
