+++ datadog_checks_dev/setup.py
@@ -80,6 +80,7 @@ setup(
             'datamodel-code-generator~=0.11.4',
             'docker-compose>=1.25',
             'in-toto>=0.4.2',
+            'jsonpointer',
             'jsonschema',
             'markdown',
             'orjson',
